Adding tags to /metadata endpoint
